test(BACKEND-ROCM): give the FusedNormRope case's eps the type the field has
`RmsNormArgs::eps` is a `float`. The case declared its constant `double` and
assigned it, which is a narrowing conversion the compiler is free to warn about
under a stricter flag set than this tree carries today, and which states the
wrong thing about the value either way: the number that reaches the kernel is
the float, so that is the number the test should name.

No behaviour changes -- 1e-6 is exactly representable either way.

FOLLOWING_AGENTS_PROTOCOL

Following-Agents-Protocol: true
AI-Assisted: true

+++ tests/vt/test_backend_cross_device.cpp
@@ -1933,7 +1933,7 @@ TEST_CASE("FusedNormRope matches the CPU oracle within NMSE <= 5e-4, both styles
   // block, so the strided reduction loop's tail is exercised rather than
   // divided away.
   constexpr int64_t kTokens = 13, kOff = 130, kRot = 16, kMaxPos = 64;
-  constexpr double kEps = 1e-6;
+  constexpr float kEps = 1e-6f;
 
   const std::vector<float> x0 = RandomVec(kTokens * (kOff + kRot), 9101);
   const std::vector<float> w0 = RandomVec(kOff, 9102, -1.0f, 1.0f);
